requirements.txt
tensorflow
scikit-learn
matplotlib
pandas
pyyaml
